crowd-counting-app/backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SocketIO==5.3.4
python-socketio==5.9.0
eventlet==0.33.3
opencv-python==4.8.1.78
numpy==1.24.3
scipy==1.11.3
tensorflow==2.13.0
Pillow==10.0.0
matplotlib==3.7.2
Werkzeug==2.3.7